requirements.txt
pillow==6.1.0
numpy==1.16.4
scipy==1.3.0
matplotlib==3.1.0
torchvision
polyaxon-client==0.5.6
future